// api/src/common/enums/TableName.ts
export enum TableName{
    ProductItem = 'product_item',
    ProductCategory = 'product_category',
    ProductCategoryToProduct = 'product_category_to_product',
    ProductAttribute = 'product_attribute',
    ProductAttributeToProduct = 'product_attribute_to_product',
    ProductGroup = 'product_group',
    ProductGroupToProduct = 'product_group_to_product',
}